+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ## [Unreleased]
 
+## [6.20.0] 2025-11-25
+
 - Prompts to disable TLS for certificate issues
+- Live DevOps Pipeline:
+  - Disables PR button when no PR exists or Git not connected
+  - Improve labels when the PR is not created yet
+- Git integration: look specifically for remote "origin"
 
 ## [6.19.0] 2025-11-25